@@ -10,8 +10,10 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.0",
         "ext-curl": "*",
+        "ext-fileinfo": "*",
+        "ext-mbstring": "*",
         "ext-soap": "*",
         "zbateson/mail-mime-parser": "^2.2"
     },
